writer/ihex: add writer for I32HEX
Introduce the binary writer module and implement
the very first format: I32HEX. The writer uses
the 00, 01 and 04 record in order to offer support
for 32-bit addressing, in case one has to deal with
bigger flash memory.

By adding the writer module, we can now perform the
complete cycle of reading a binary file, disassembling it,
working with the resulting instructions, assembling them
back to opcodes and finally writing everything back to
a binary file.

Testing of this new feature is a WIP.

Signed-off-by: Milo-D <milod2048@gmail.com>